Move icon-picker to separate package, remove Translate.component and implement the translation directly in data-table and icon-picker
