+++ packages/core/src/index.js
@@ -21,9 +21,6 @@ export { default as DropDown } from './form-fields/DropDown.component'
 export { default as MultiToggle } from './form-fields/MultiToggle'
 export { default as TextField } from './form-fields/TextField'
 export { default as Heading } from './headings/Heading.component'
-export { default as CurrentIcon } from './icon-picker/CurrentIcon.component'
-export { default as IconOption } from './icon-picker/IconOption.component'
-export { default as IconPicker } from './icon-picker/IconPicker.component'
 export { default as Column } from './layout/Column.component'
 export { default as Row } from './layout/Row.component'
 export { default as SinglePanel } from './layout/SinglePanel.component'
