+++ packages/icon-picker/package.json
@@ -1,0 +1,41 @@
+{
+    "name": "d2-ui-icon-picker",
+    "version": "1.0.2",
+    "description": "DHIS2 component for selecting icon",
+    "main": "./build/cjs/index.js",
+    "module": "./build/es/index.js",
+    "files": [
+        "build"
+    ],
+    "license": "BSD-3-Clause",
+    "peerDependencies": {
+        "d2": "^29.1.1",
+        "react": "^16.0.0",
+        "react-dom": "^16.0.0"
+    },
+    "author": "Mark Polak",
+    "contributors": [
+        "Nicolay Ramm",
+        "Kjetil Svalestuen <kjetil@dhis2.org>",
+        "Viktor Varland <viktor@dhis2.org>",
+        "Edoardo Sabadelli <edoardo@dhis2.org>"
+    ],
+    "scripts": {
+        "prebuild": "rimraf ./build/*",
+        "build": "npm run build:commonjs && npm run build:es && npm run build:umd",
+        "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir build/cjs --ignore spec.js",
+        "build:es": "cross-env BABEL_ENV=es babel src --out-dir build/es --ignore spec.js",
+        "build:umd": "cross-env && webpack -p",
+        "watch": "npm run build:es --  --watch"
+    },
+    "dependencies": {
+        "d2-ui": "^29.0.5",
+        "material-ui": "^0.19.3",
+        "prop-types": "^15.5.10"
+    },
+    "jest": {
+        "transformIgnorePatterns": [
+            "node_modules/(?!react-native|react-navigation)/"
+        ]
+    }
+}
