+++ packages/icon-picker/src/IconOption.component.js
