+++ packages/icon-picker/src/IconPicker.component.js
