+++ packages/icon-picker/src/__tests__/CurrentIcon.component.spec.js
