+++ packages/icon-picker/src/__tests__/IconOption.component.spec.js
