+++ packages/icon-picker/src/__tests__/IconPicker.component.spec.js
