+++ packages/icon-picker/src/index.js
@@ -1,0 +1,3 @@
+import IconPicker from './IconPicker.component';
+
+export default IconPicker;
