+++ packages/translation-dialog/src/index.js
@@ -1,5 +1,3 @@
 import TranslationDialog from './TranslationDialog.component';
 
-export { default as Translate } from './Translate.component';
-
 export default TranslationDialog;
